chore(docs): remove duplicate Wing SDK spec (#2926)
There were two copies of the Wing SDK spec, so I removed the duplicate.

I also cleaned up some internal naming in the SDK to consolidate some of the code around the API name change from `addConsumer` to `setConsumer`. (This is sort of a breaking change, but in practice I doubt anyone's using the interface name directly).

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.