chore: run unit tests in parallel with build/e2e and fix e2e sharding (#2944)
Time Save: ~17 minutes
Time Save (pr/e2e-full): ~30 minutes

This is a low-effort change, but saves us quite a bit of time.
This adds quite a bit of duplication due to the fact that testing in a new job effectively requires us to redo a lot of work, but at the cost of wasted workflow time we gain overall speed. This could be possible to avoid by saving the entire build steps as an asset, but that would be multiple GBs of data so I'm not sure it's worth it. 
It would also be solved by fully figuring out Nx caching, but tbh I think that'll be a nightmare.

Additionally, it turns out that the test sharding for e2e tests weren't working. So we were just running all tests twice 😭 . Fixing that shaved another ~7 minutes from non-full e2e load.

Misc Change: Decided also to remove the auto-approve workflow, since we're not using it and it adds a useless "skipped" status to all PRs.

Closes https://github.com/winglang/wing/issues/2820

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.