chore: use PROJEN_BUMP_VERSION for versioning (#2993)
Another followup to #2992

Switched back to using PROJEN_BUMP_VERSION, which is what the original script used. The issue with the last change probably was that we were trying to regenerate the version/changelog and it was failing. Using the existing variable should work and it's faster anyways.

Also explicitly added a little log, which shouldn't hurt.