fix: capture diagnostics missing from language server (#2996)
Certain errors were not available in the lsp because we didn't run the jsify phase. This PR changes that and also avoids many panics. 

So far it seems the perf overhead of running jsify in the lsp is minimal. On my PC, it adds between 1ms and 8ms. This time scales with classes/inflight/externs.

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.