fix(compiler): panic when trying to import JSII constructs with no scope/id args (#3057)
See #2934 
Importing a construct from JSII that doesn't have a `scope`/`id` arg pair as the first constructor arguments panicked the compiler. We now detect such (unconventional) constructs and mark them as so. This makes them non compatible with our `as`/`in` keywords for explicit scope/id control, but they can still be used by passing the correct (non standard) arguments to them.
This resolves importing the cdktf library because its Backend constructs are an example of such non-standard constructs. 

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.