feat(docs): split API Reference by submodule (#3113)
Split up the API reference so each module in the standard library (`cloud`, `http`, `util`, etc.) has its own dedicated page. Besides making it easier to view, hopefully will make it easier for headings in these pages to get indexed and (crossing our fingers 🤞) will address https://github.com/winglang/wing/issues/2710.

<img width="230" alt="Screenshot 2023-06-27 at 6 30 04 PM" src="https://github.com/winglang/wing/assets/5008987/89316047-df92-437a-9bff-aec2fa5c0655">

It's not quite perfect (if there is a link to a resource in another subpage, the link just won't do anything) but it's a step in the right direction.

## Checklist

- [ ] Title matches [Winglang's style guide](https://docs.winglang.io/contributing/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.