fix(vscode): static completions for `Json.` wrong/missing after other expressions (#3137)
I believe this is a regression from https://github.com/winglang/wing/pull/3128. See the new test cases for the issues this covers. This issue is kinda specific to Json because it has its own tree-sitter node for a single identifier unlike others.

Also renamed a symbol in the completion visitor because `location` was a bad/wrong name.

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributing/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.