feat: expression lifting
## TODO

- [ ] Still failing tests
- [ ] Determine how to pass lifting tokens to super classes

Refactor the lifting system to work on expressions instead of references and merge lifting and capturing into a unified model.

*Lifting* is the concept of referencing a preflight expression from an inflight scope, while *capturing* is the concept of referencing an object defined in a parent scope (albeit not another phase).

The mechanism works like this: we now have another "lifting" compilation phase which runs immediately after type checking. This is a transform ("fold") which does the following:

1. Lifting: for all inflight methods, it scans all the expressions. If it finds and expression marked as a preflight expression, it transforms it into an `Expr::Lift` which is a placeholder that can be used later to replace it with a lifted token.
2. Capturing: for each reference (both type refs and object refs), it checks looks up in which symbol environment it is defined. It then checks if that environment is a parent of the current method environment. If it is, is transforms it into a `Reference::Capture` to indicate downstream that this is a captured reference.

Now we move to the jsification phase. When the jsifier hits a *lifted expression* or a *captured reference*, to jsifies it as if it was in a preflight scope and registers it in a centralized token registry maintained at the class level during jsification. This registry maintains a mapping between *inflight tokens* and *preflight expressions* (both for lifting and for capturing purposes).

The token registry is consulted when the class is jsified in order to emit the code that lifts expressions from preflight to inflight (both at the object level and at the type level).

## Fixes

* #1878 is fixed because now we are able to lift complete expressions. So `bucket.bucketArn` is a simple string and there is no need to lift the bucket.

This is the 3rd attempt at this refactor, and hopefully the last for now. It supersedes #3125 and #2909.