fix: missing access information for inflight initializers (take II) (#3316)
Fixes #3301 by adding a synthetic lift between the `$inflight_init` method and all lifted objects and also synthetically always binding objects to the init method.

This is in order to be able to create all inflight objects regardless of the actual access of the consumer of the inflight object.

This is take II for #3302, which surfaced a latent bug in which inflight fields were accidentally lifted.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://www.winglang.io/terms-and-policies/contribution-license.html)*.