fix: labeling during self mutation not working for pull requests from forks (#3392)
Labeling during self mutation was added as a cautionary method but seems like it is not working for PRs from forks. This is because github context does not return pull request data (at least in this case) and so there is no way to form the proper url for rest api call using just github context. Instead retrieve the PR number with a rest api call and use it. Also change the labeling to after mutation commit.
Modified the step to use script module instead of run as per @MarkMcCulloh's suggestion.
fixes #3391.
## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://www.winglang.io/terms-and-policies/contribution-license.html)*.