feat(compiler): multi-file Wing applications (#3512)
Introduce a basic form of multi-file Wing projects through the `bring` keyword. To use classes, interfaces, enums, or structs from another Wing file, bring a file with a path that ends in `.w`, and assign it to an alias:

```js
bring "./other-file.w" as shared;

new shared.Store();
```

This initial implementation follows a (naive) strategy of recursively instantiating the parser to create a large AST. This has some limitations -- for example it's not possible for two Wing files to reuse the same third Wing file -- but this still gives users a lot more refactoring power for building Wing apps. #3511 is the main follow-up issue for addressing this.

Closes #476

Follow up:
- https://github.com/winglang/wing/issues/3509
- https://github.com/winglang/wing/issues/3511

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.