fix(docs): enhancing the getting-started app on the docs (#3551)
This PR is to enhance the hello.w getting-started app in the docs. Initially adding a counter to the app, but also an API is on planning - so don't merge it yet...
I'm open to wording changes (as it's wing's front door), and generally open to any feedback 😄

Tested on the simulator, and also by running `terraform apply` - but without really applying it and testing on AWS console. Will test it too
Enhancement issue #3541

## Checklist

- [ ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.

(cherry picked from commit 255836df380dd01718c5664258fb07761d958b9d)