fix(docs): more enhancing to the getting-started app on the docs (#3576)
This PR is to continue the enhancement of the hello.w getting-started app in the docs. 
I'm open to wording changes here too, and generally open to any feedback 😄

Tested on the simulator, and also by running terraform apply and seeing on AWS console.
Enhancement issue https://github.com/winglang/wing/issues/3541

## Checklist

- [ ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.