feat(compiler): variadic arguments (#3643)
This PR adds the possibility of using variadic arguments

Unfortunately, I had to add the `Clone` attribute to many `enum` and `struct` to be able to create a list of arguments compatible with the requirements for verifying the typing of the variadic argument.

I needed to implement `PartialEq` for `Type`, `TypeRef`, `Class`, `Interface`, `Struct`, and `Enum`.
I'm not sure if it was done in the best way, but it fulfills this specific case.

- [x] Validates if the last argument (variadic) of the function accepts multiple elements.
- [x] Validates if the only argument (variadic) of the function accepts multiple elements.
- [x] Validates if all types of the variadic argument are the same.
- [ ] Accept range as a variadic argument.

Closes #125 
Closes #397

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.