fix: cannot extend non-wing construct
When trying to bring `cdk8s` and extend `Chart`, the call to `this._addInflightOps()` fails because non-wing constructs don't have this method.

The fix is to use `?.()` (e.g. `this._addInflightOps?.()`) in order to chain the undefined.

Added a test which verifies that cdk8s can be brought and used.

Fixes https://github.com/winglang/wing/issues/3717