fix(compiler): missing type check for partially inferred function (#3719)
Fixes #3714

When making an inference in type validation, the inferred type was originally thrown away. This change instead unwraps it if needed since updating the original typeref itself won't happen til later. It's worth noting the error message will be missing information for deeply inferred types (like the return type).

Also it turns out this regression was noticeable in the snapshots before, but not noticed. So the snapshot update included 2 changes.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.