chore(console): stop running playwright tests during build (#3722)
Renames the `@wingconsole/app#test` script to `@wingconsole/app#test:playwright`, so it doesn't run with the root scripts `build` or `test`. The problem is that in order to run Playwright tests, users need to run `playwright install` or `playwright install --with-deps` first, and may require sudo permissions.

Additionally, moves the playwright tests from `tests/` to `test/`, which is the standard directory for both playwright and our turbo configuration.

Supersedes https://github.com/winglang/wing/pull/3721.