make assingment `lvalue` a reference instead of expression
This reverts a change done in #3213 where we added phases to all expressions. In practice it doesn't seem the assignment target itself needs a phase other than the phase already assingned to the inner expression in the reference and the phase assigned to the variable itself.
This also removed a duplicate error message when trying to access `this` from a static function due to a double resolving caused by resolving the entire expression and then the inner expression.