chore(compiler): remove object freezing (#3733)
Today, object freezing in Wing doesn't provide much value to users. The compiler emits `Object.freeze()` in lots of the JS code for the promise of runtime safety, but to date it isn't a capability we test.

To that end, I'm proposing to remove object freezing until we have stronger use cases for this. By removing any details on freezing, we're also able to simplify Wing's type model and reduce the code we're emitting in both preflight and inflight.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [x] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.